+++ pom.xml
@@ -60,8 +60,8 @@
         <version>3.3</version>
         <configuration>
           <useIncrementalCompilation>false</useIncrementalCompilation>
-          <source>1.7</source>
-          <target>1.7</target>
+          <source>1.8</source>
+          <target>1.8</target>
           <compilerArgs>
             <compilerArg>-Xlint:all</compilerArg>
           </compilerArgs>
