+++ processor/src/main/java/io/norberg/automatter/processor/BuildStatements.java
@@ -1,0 +1,11 @@
+package io.norberg.automatter.processor;
+
+public class BuildStatements {
+  Iterable<Statement> statements;
+  String parameter;
+
+  public BuildStatements(Iterable<Statement> statements, String parameter) {
+    this.statements = statements;
+    this.parameter = parameter;
+  }
+}
