+++ processor/src/main/java/io/norberg/automatter/processor/Statement.java
@@ -1,0 +1,11 @@
+package io.norberg.automatter.processor;
+
+public class Statement {
+  String statement;
+  Object[] args;
+
+  public Statement(String statement, Object... args) {
+    this.statement = statement;
+    this.args = args;
+  }
+}
